refactor(webui): split settings view components
